refactor(app): S10–S11 store 服务层拆分与壳层瘦身
将 imageStore/workspaceStore 的上传、变更、列表与本地 FS 逻辑抽到服务模块，App/LibraryPage 编排迁入 hooks，删除 pages/ 并将 LibraryPage 归入 features/library。
